coxir v0.1.0 Coxir

When used, this module defines a start_link/0 callback responsible for starting a Consumer process. This process then consumes events from the GenStage and then dispatches them onto the handling module.

Essentially that-is Consumer is provided with the module’s reference the start_link callback was defined on, and thus is then able to call the custom handling methods that one may define based on their needs.

Additionally, aliases to all the structures as well as to other crucial modules are included when used.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.