Coxir.API (coxir v2.0.0)

Entry-point to the Discord REST API.

Types

options()

Specs

options() :: [
  as: Coxir.Gateway.gateway() | none(),
  token: Coxir.Token.t() | none()
]

The options that can be passed to perform/4.

If the :as option is present, the token of the given gateway will be used.

If no token is provided, one is expected to be configured as :token under the :coxir app.

result()

Specs

result() :: :ok | {:ok, map()} | {:error, Coxir.API.Error.t()}

The possible outcomes of perform/4.

Functions

delete(path, options)

Specs

delete(Tesla.Env.url(), options()) :: result()

Delegates to perform/4 with method set to :delete.

get(path, options)

Specs

get(Tesla.Env.url(), options()) :: result()

Delegates to perform/4 with method set to :get.

patch(path, body, options)

Specs

patch(Tesla.Env.url(), Tesla.Env.body(), options()) :: result()

Delegates to perform/4 with method set to :patch.

perform(method, path, body \\ nil, options)

Specs

perform(Tesla.Env.method(), Tesla.Env.url(), Tesla.Env.body(), options()) ::
  result()

Performs a request to the API.

post(path, body \\ %{}, options)

Specs

post(Tesla.Env.url(), Tesla.Env.body(), options()) :: result()

Delegates to perform/4 with method set to :post.

put(path, body \\ %{}, options)

Specs

put(Tesla.Env.url(), Tesla.Env.body(), options()) :: result()

Delegates to perform/4 with method set to :put.