Coxir.Channel (coxir v2.0.0)

Represents a Discord channel.

Types

application_id()

Specs

application_id() :: Coxir.Model.Snowflake.t() | nil

Application id of the group DM creator if it is bot-created.

bitrate()

Specs

bitrate() :: non_neg_integer() | nil

The bitrate (in bits) of the voice channel.

create_params()

Specs

create_params() :: Enum.t() | create_params_dm() | create_params_guild()

The parameters that can be passed to create/2.

create_params_dm()

Specs

create_params_dm() :: %{recipient_id: recipient_id()}

Parameters when creating a DM channel.

create_params_guild()

Specs

create_params_guild() :: %{
  :guild_id => guild_id(),
  :name => name(),
  optional(:type) => type(),
  optional(:topic) => topic(),
  optional(:bitrate) => bitrate(),
  optional(:user_limit) => user_limit(),
  optional(:rate_limit_per_user) => rate_limit_per_user(),
  optional(:position) => position(),
  optional(:permission_overwrites) => permission_overwrites(),
  optional(:parent_id) => parent_id(),
  optional(:nsfw) => nsfw()
}

Parameters when creating a guild channel.

guild()

Specs

guild() ::
  Ecto.Association.NotLoaded.t() | Coxir.Guild.t() | nil | Coxir.API.Error.t()

The guild the channel belongs to.

Needs to be preloaded via preload/3.

guild_id()

Specs

guild_id() :: Coxir.Model.Snowflake.t() | nil

The id of the guild the channel belongs to.

icon()

Specs

icon() :: String.t() | nil

The icon hash of the channel.

id()

Specs

id() :: Coxir.Model.Snowflake.t()

The id of the channel.

invites()

Specs

invites() ::
  Ecto.Association.NotLoaded.t() | [Coxir.Invite.t()] | Coxir.API.Error.t()

Invites for the channel.

Needs to be preloaded via preload/3.

key()

Specs

key() :: id()

The coxir key of a channel.

name()

Specs

name() :: String.t() | nil

The name of the channel.

nsfw()

Specs

nsfw() :: boolean() | nil

Whether the channel is nsfw.

owner()

Specs

owner() ::
  Ecto.Association.NotLoaded.t() | Coxir.User.t() | nil | Coxir.API.Error.t()

The creator of the group DM or thread.

Needs to be preloaded via preload/3.

owner_id()

Specs

owner_id() :: Coxir.Model.Snowflake.t() | nil

The id of the creator of the group DM or thread.

parent()

Specs

parent() :: Ecto.Association.NotLoaded.t() | t() | nil | Coxir.API.Error.t()

The parent category for guild channels. The belonging channel for threads.

Needs to be preloaded via preload/3.

parent_id()

Specs

parent_id() :: Coxir.Model.Snowflake.t() | nil

The id of the parent category for guild channels. The id of the belonging channel for threads.

permission_overwrites()

Specs

permission_overwrites() ::
  Ecto.Association.NotLoaded.t() | [Coxir.Overwrite.t()] | Coxir.API.Error.t()

Permission overwrites for members and roles.

Needs to be preloaded via preload/3.

pinned_messages()

Specs

pinned_messages() :: [Coxir.Message.t()] | Coxir.API.Error.t()

The pinned messages in the channel.

Needs to be preloaded via preload/3.

position()

Specs

position() :: non_neg_integer() | nil

Sorting position of the channel.

rate_limit_per_user()

Specs

rate_limit_per_user() :: non_neg_integer() | nil

Amount of seconds a user has to wait before sending another message.

recipient_id()

Specs

recipient_id() :: Coxir.Model.Snowflake.t()

The id of the recipient user of the DM.

recipients()

Specs

recipients() :: [Coxir.User.t()] | nil

The recipients of the DM.

Needs to be preloaded via preload/3.

rtc_region()

Specs

rtc_region() :: String.t() | nil

The voice region id for the voice channel, nil when automatic.

t()

Specs

t() :: %Coxir.Channel{
  application_id: application_id(),
  bitrate: bitrate(),
  guild: guild(),
  guild_id: guild_id(),
  icon: icon(),
  id: id(),
  invites: invites(),
  name: name(),
  nsfw: nsfw(),
  owner: owner(),
  owner_id: owner_id(),
  parent: parent(),
  parent_id: parent_id(),
  permission_overwrites: permission_overwrites(),
  pinned_messages: pinned_messages(),
  position: position(),
  rate_limit_per_user: rate_limit_per_user(),
  recipients: recipients(),
  rtc_region: rtc_region(),
  topic: topic(),
  type: type(),
  user_limit: user_limit(),
  video_quality_mode: video_quality_mode(),
  voice_states: voice_states(),
  webhooks: webhooks()
}

The struct for a channel.

topic()

Specs

topic() :: String.t() | nil

The topic of the channel.

type()

Specs

type() :: non_neg_integer() | nil

The type of the channel. For a list of types have a look here.

update_params()

Specs

update_params() :: Enum.t() | update_params_guild()

The parameters that can be passed to update/2.

update_params_guild()

Specs

update_params_guild() :: %{
  optional(:type) => type(),
  optional(:topic) => topic(),
  optional(:bitrate) => bitrate(),
  optional(:user_limit) => user_limit(),
  optional(:rate_limit_per_user) => rate_limit_per_user(),
  optional(:position) => position(),
  optional(:permission_overwrites) => permission_overwrites(),
  optional(:parent_id) => parent_id(),
  optional(:nsfw) => nsfw()
}

Parameters when updating a guild channel.

user_limit()

Specs

user_limit() :: non_neg_integer() | nil

The user limit of the voice channel.

video_quality_mode()

Specs

video_quality_mode() :: non_neg_integer() | nil

The camera video quality mode of the voice channel.

voice_states()

Specs

voice_states() :: Ecto.Association.NotLoaded.t() | [Coxir.VoiceState.t()]

Active voice states for the voice channel.

Needs to be preloaded via preload/3.

webhooks()

Specs

webhooks() ::
  Ecto.Association.NotLoaded.t() | [Coxir.Webhook.t()] | Coxir.API.Error.t()

Webhooks configured for the channel.

Needs to be preloaded via preload/3.

Functions

bulk_delete_messages(channel, messages, options \\ [])

Specs

bulk_delete_messages(
  t(),
  [Coxir.Message.t()] | [Coxir.Model.Snowflake.t()],
  Coxir.Model.Loader.options()
) :: Coxir.Model.Loader.result()

Deletes messages in bulk from a given channel.

This only works for messages not older than 2 weeks.

create(params, options \\ [])

Specs

create(create_params(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Callback implementation for Coxir.Model.create/2.

create_invite(channel, params \\ %{}, options \\ [])

Specs

create_invite(t(), Enum.t(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Delegates to Coxir.Invite.create/2.

create_overwrite(channel, params, options \\ [])

Specs

create_overwrite(t(), Enum.t(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Delegates to Coxir.Overwrite.create/2.

create_webhook(channel, params, options \\ [])

Specs

create_webhook(t(), Enum.t(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Delegates to Coxir.Webhook.create/2.

delete(struct, options \\ [])

Callback implementation for Coxir.Model.delete/2.

get(key, options \\ [])

Callback implementation for Coxir.Model.get/2.

get!(key, options \\ [])

Callback implementation for Coxir.Model.get!/2.

get_message(channel, message_id, options \\ [])

Specs

get_message(t(), Coxir.Model.Snowflake.t(), Coxir.Model.Loader.options()) ::
  Coxir.Message.t() | Coxir.API.Error.t()

Delegates to Coxir.Message.get/2.

preload(struct, preloads, options \\ [])

Callback implementation for Coxir.Model.preload/3.

preload!(struct, preloads, options \\ [])

Callback implementation for Coxir.Model.preload!/3.

send_message(channel, params, options \\ [])

Specs

send_message(t(), Enum.t(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Delegates to Coxir.Message.create/2.

start_typing(channel, options \\ [])

Specs

start_typing(t(), Coxir.Model.Loader.options()) :: Coxir.Model.Loader.result()

Triggers the typing indicator on a given channel.

update(struct, params, options \\ [])

Specs

update(t(), update_params(), Coxir.Model.Loader.options()) ::
  Coxir.Model.Loader.result()

Callback implementation for Coxir.Model.update/3.