Coxir.Gateway (coxir v2.0.0)

Supervises the components necessary to interact with Discord's gateway.

Using the module

The module can be used with Kernel.use/2 to transform the calling module into a gateway.

Check out the __using__/1 macro to see what this does exactly.

Types

config()

Specs

config() :: [
  token: Coxir.Token.t() | none(),
  intents: Coxir.Gateway.Intents.intents() | :non_privileged,
  shard_count: non_neg_integer() | none(),
  sharder: Coxir.Sharder.t() | Coxir.Sharder.Default,
  handler: Coxir.Gateway.Handler.handler()
]

The configuration that must be passed to start_link/2.

If no token is provided, one is expected to be configured as :token under the :coxir app.

If no shard_count is provided, the value suggested by Discord will be used.

gateway()

Specs

gateway() :: Supervisor.supervisor()

A gateway process.

Functions

__using__(options)
(macro)

Defines functions in order to transform the calling module into a gateway.

Defines child_spec/1 and start_link/0 so that the module can be added to a supervisor.

Note that the gateway/0 process that the second function starts is named after the module.

Defines get_user_id/0 which delegates to get_user_id/1.

Defines update_presence/1 which delegates to update_presence/2.

Defines update_presence/2 which delegates to update_presence/3.

Requires the Coxir.Gateway.Handler behaviour to be implemented for handling events.

Custom configuration can be given for this module by configuring it under the :coxir app.

get_producer(gateway)

Specs

get_producer(gateway()) :: Coxir.Gateway.Producer.producer()

Returns the Coxir.Gateway.Producer.producer/0 process of a given gateway.

get_shard(gateway, index)

Specs

get_shard(gateway(), Coxir.Channel.t() | Coxir.Guild.t() | non_neg_integer()) ::
  Coxir.Gateway.Session.session()

Returns the session process for a given channel, guild or specific shard.

get_token(gateway)

Specs

get_token(gateway()) :: Coxir.Token.t()

Returns the token configured for the given gateway.

get_user_id(gateway)

Specs

get_user_id(gateway()) :: Coxir.Model.Snowflake.t()

Returns the id of the user the given gateway is running for.

request_guild_members(gateway, guild, params)

Specs

request_guild_members(gateway(), Coxir.Guild.t(), Enum.t()) :: :ok

Requests members for a given guild.

The possible parameters are the fields of Coxir.Gateway.Payload.RequestGuildMembers.t/0.

start_link(config, options \\ [])

Specs

start_link(config(), [Supervisor.option() | Supervisor.init_option()]) ::
  Supervisor.on_start()

Starts a gateway with the given configuration and options.

update_presence(gateway, params)

Specs

update_presence(gateway(), Enum.t()) :: :ok

Calls update_presence/3 on all the shards of a given gateway.

update_presence(gateway, where, params)

Specs

update_presence(
  gateway(),
  Coxir.Channel.t() | Coxir.Guild.t() | non_neg_integer(),
  Enum.t()
) :: :ok

Updates the presence on a given channel, guild or specific shard.

The possible parameters are the fields of Coxir.Gateway.Payload.UpdatePresence.t/0.