Coxir.Model.Loader (coxir v2.0.0)

Implements functionality for Coxir.Model.model/0 implementations.

Types

loader_options()

Specs

loader_options() :: [
  force: boolean() | false,
  storage: boolean() | true,
  fetch: boolean() | true
]

The options specific to this module.

The force option specifies whether an association must be reloaded.

The storage option specifies whether the storage should be hit.

The fetch option specifies whether the API should be hit.

options()

Specs

options() :: loader_options() | Coxir.API.options()

The options that can be passed to the functions.

preloads()

Specs

preloads() :: atom() | [atom()] | [{atom(), preloads()}]

The format of the preloads argument of preload/3.

result()

Specs

result() :: {:ok, Coxir.Model.instance()} | Coxir.API.result()

The possible outcomes of a write operation.

Functions

create(model, params, options)

Specs

create(Coxir.Model.model(), Enum.t(), options()) :: result()

Attempts to create a Coxir.Model.instance/0 through the API.

The created entity will be passed through load/2 on success.

delete(struct, options)

Specs

delete(Coxir.Model.instance(), options()) :: result()

Attempts to delete a Coxir.Model.instance/0 through the API.

The deleted entity will be passed through unload/1 on success.

get(model, key, options)

Specs

get(Coxir.Model.model(), Coxir.Model.key(), options()) ::
  Coxir.Model.instance() | Coxir.API.Error.t()

Attempts to get a Coxir.Model.instance/0 from the storage and/or the API.

If fetched from the API, the instance will be passed through load/2 on success.

get!(model, key, options)

Specs

get!(Coxir.Model.model(), Coxir.Model.key(), options()) ::
  Coxir.Model.instance()

Same as get/3 but raises on error.

load(model, objects)

Specs

load(Coxir.Model.model(), [map()]) :: [Coxir.Model.instance()]

load(Coxir.Model.model(), map()) :: Coxir.Model.instance()

Loads Coxir.Model.instance/0 from a map/0 object.

This function also stores the models configured as storable.

preload(structs, preloads, options)

Specs

preload([Coxir.Model.instance()], preloads(), options()) :: [
  Coxir.Model.instance()
]

preload(Coxir.Model.instance(), preloads(), options()) :: Coxir.Model.instance()

Attempts to load the associations for a Coxir.Model.instance/0.

If fetched from the API, the associations will be passed through load/2 on success.

preload!(structs, preloads, options)

Specs

preload!([Coxir.Model.instance()], preloads(), options()) :: [
  Coxir.Model.instance()
]

preload!(Coxir.Model.instance(), preloads(), options()) ::
  Coxir.Model.instance()

Same as preload/3 but raises on error.

unload(struct)

Specs

unload(Coxir.Model.instance()) :: :ok

Removes a Coxir.Model.instance/0 from the storage when of a storable model.

update(struct, params, options)

Specs

update(Coxir.Model.instance(), Enum.t(), options()) :: result()

Attempts to update a Coxir.Model.instance/0 through the API.

The updated entity will be passed through load/2 on success.