Coxir.Player behaviour (coxir v2.0.0)

Handles the audio sent through voice.

Types

init_argument()

Specs

init_argument() :: {playable(), options()}

The argument that is passed to child_spec/1.

options()

Specs

options() :: keyword()

Specifies the options that can be passed to the t/0 in use.

playable()

Specs

playable() :: term()

Specifies what can be played with the t/0 in use.

player()

Specs

player() :: pid()

A player process.

t()

Specs

t() :: module()

A module that implements the behaviour.

Callbacks

child_spec(init_argument)

Specs

child_spec(init_argument()) :: Supervisor.child_spec()

Must return a child specification from a init_argument/0.

invalidate(player)

Specs

invalidate(player()) :: :ok

Called when the connection to the Discord voice channel is lost.

This invalidates the previously received Coxir.Voice.Audio.t/0 struct.

Thus the player should stop sending audio until ready/2 is called again.

pause(player)

Specs

pause(player()) :: :ok

Called to pause audio playback.

playing?(player)

Specs

playing?(player()) :: boolean()

Called to check whether audio playback is paused.

ready(player, arg2)

Specs

ready(player(), Coxir.Voice.Audio.t()) :: :ok

Called when the connection to the Discord voice channel is ready.

The received Coxir.Voice.Audio.t/0 struct can be used to send audio.

resume(player)

Specs

resume(player()) :: :ok

Called to resume audio playback.