Coxir.Storage behaviour (coxir v2.0.0)

Handles how models are cached.

Configuration

A list of models to be stored can be configured as :storable under the :coxir app.

If no such list is configured, all models hardcoded as storable will be stored by default.

A custom storage adapter can be configured as :adapter under the :coxir app.

If no custom adapter is configured, the default Coxir.Storage.Default will be used.

Callbacks

all(arg1)

Specs

all(Coxir.Model.model()) :: [Coxir.Model.instance()]

all_by(arg1, keyword)

Specs

all_by(Coxir.Model.model(), keyword()) :: [Coxir.Model.instance()]

child_spec(term)

Specs

child_spec(term()) :: Supervisor.child_spec()

delete(arg1, arg2)

Specs

delete(Coxir.Model.model(), Coxir.Model.key()) :: :ok

delete_by(arg1, keyword)

Specs

delete_by(Coxir.Model.model(), keyword()) :: :ok

get(arg1, arg2)

Specs

get(Coxir.Model.model(), Coxir.Model.key()) :: Coxir.Model.instance() | nil

get_by(arg1, keyword)

Specs

get_by(Coxir.Model.model(), keyword()) :: Coxir.Model.instance() | nil

put(arg1)

Specs

put(Coxir.Model.instance()) :: Coxir.Model.instance()