coxir v0.1.0 Coxir.Struct.Channel

Defines methods used to interact with Discord channels.

Refer to this for a list of fields and a broader documentation.

In addition, the following fields are also embedded.

owner - an user object

Types

channel()

channel() :: String.t() | map()

Functions

bulk_delete_messages(channel, messages)

bulk_delete_messages(channel(), list()) :: :ok | map()

Deletes multiple messages from a given channel.

Returns the atom :ok upon success or a map containing error information.

create_invite(term, params \\ %{})

create_invite(channel(), Enum.t()) :: map()

Creates an invite for a given channel.

Returns an invite object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

max_uses - max number of uses (0 if unlimited)

max_age - duration in seconds before expiry (0 if never)

temporary - whether this invite only grants temporary membership

unique - whether not to try to reuse a similar invite

Refer to this for a broader explanation on the fields and their defaults.

create_permission(channel, overwrite, params)

create_permission(channel(), String.t(), Enum.t()) :: map()

Creates a permission overwrite for a given channel.

Refer to Coxir.Struct.Overwrite.edit/2 for more information.

create_webhook(channel, params)

create_webhook(channel(), Enum.t()) :: map()

Creates a webhook for a given channel.

Returns a webhook object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - name of the webhook (2-23 characters)

avatar - image for the default webhook avatar

delete(channel)

delete(channel()) :: map()

Deletes a given channel.

Returns a channel object upon success or a map containing error information.

edit(channel, params)

edit(channel(), Enum.t()) :: map()

Modifies a given channel.

Returns a channel object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - channel name (2-100 characters)

topic - channel topic (up to 1024 characters)

nsfw - whether the channel is NSFW

position - the position in the left-hand listing

bitrate - the bitrate in bits of the voice channel

user_limit - the user limit of the voice channel

permission_overwrites - channel or category-specific permissions

parent_id - id of the new parent category

Refer to this for a broader explanation on the fields and their defaults.

get(id)

get(String.t()) :: map() | nil

Fetches a cached channel object.

Returns an object if found and nil otherwise.

get_invites(channel)

get_invites(channel()) :: list() | map()

Fetches the invites from a given channel.

Returns a list of invite objects upon success or a map containing error information.

get_message(channel, message)

get_message(channel(), String.t()) :: map()

Fetches a message from a given channel.

Returns a message object upon success or a map containing error information.

get_pinned_messages(channel)

get_pinned_messages(channel()) :: list() | map()

Fetches the pinned messages from a given channel.

Returns a list of message objects upon success or a map containing error information.

get_voice_members(channel)

get_voice_members(channel()) :: list() | map()

Fetches the users currently in a given voice channel.

Returns a list of user or member objects depending on whether it’s a private or a guild channel respectively or a map containing error information.

get_webhooks(channel)

get_webhooks(channel()) :: list() | map()

Fetches the webhooks from a given channel.

Returns a list of webhook objects upon success or a map containing error information.

history(term, query \\ [])

history(channel(), Keyword.t()) :: list() | map()

Fetches messages from a given channel.

Returns a list of message objects upon success or a map containing error information.

Query

Must be a keyword list with the fields listed below.

around - get messages around this message ID

before - get messages before this message ID

after - get messages after this message ID

limit - max number of messages to return

Refer to this for a broader explanation on the fields and their defaults.

select(pattern \\ %{})

select(map()) :: list()

Filters the cached channel objects.

Returns a list of matched objects.

send_message(channel, content)

send_message(channel(), String.t() | Enum.t()) :: map()

Sends a message to a given channel.

Returns a message object upon success or a map containing error information.

Content

Either a string or an enumerable with the fields listed below.

content - the message contents (up to 2000 characters)

embed - embedded rich content, refer to this

nonce - used for optimistic message sending

file - the path of the file being sent

tts - true if this is a TTS message

Refer to this for a broader explanation on the fields and their defaults.