coxir v0.8.0 Coxir.Struct.Guild

Defines methods used to interact with Discord guilds.

Refer to this for a list of fields and a broader documentation.

In addition, the following fields are also embedded.

owner - a user object

afk_channel - a channel object

embed_channel - a channel object

system_channel - a channel object

channels - list of channel objects

members - list of member objects

roles - list of role objects

Types

guild()

guild() :: String.t() | map()

Functions

add_member(guild, user, params)

add_member(guild(), String.t(), Enum.t()) :: map()

Adds a user to a given guild.

Returns a member object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

access_token - an oauth2 access token

nick - value to set the user’s nickname to

roles - list of role ids the user is assigned

mute - whether the user is muted

deaf - whether the user is deafened

Refer to this for a broader explanation on the fields and their defaults.

create_channel(guild, params)

create_channel(guild(), Enum.t()) :: map()

Creates a channel on a given guild.

Returns a channel object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - channel name (2-100 characters)

type - the type of channel

nswf - whether the channel is NSFW

bitrate - the bitrate in bits of the voice channel

user_limit - the user limit of the voice channel

permission_overwrites - channel-specific permissions

parent_id - id of the parent category

Refer to this for a broader explanation on the fields and their defaults.

create_integration(guild, params)

create_integration(guild(), Enum.t()) :: :ok | map()

Attaches an integration to a given guild.

Returns the atom :ok upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

type - the integration type

id - the integration id

create_role(guild, params \\ %{})

create_role(guild(), Enum.t()) :: map()

Creates a role on a given guild.

Returns a role object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - name of the role

color - RGB color value

permissions - bitwise of the permissions

hoist - whether the role should be displayed separately

mentionable - whether the role should be mentionable

Refer to this for a broader explanation on the fields and their defaults.

delete(guild)

delete(guild()) :: :ok | map()

Deletes a given guild.

Returns the atom :ok upon success or a map containing error information.

do_prune(guild, query)

do_prune(guild(), Keyword.t()) :: map()

Begins a prune operation for a given guild.

Returns a map with a pruned field or a map containing error information.

Query

Must be a keyword list with the fields listed below.

days - number of days to count prune for (1 or more)

edit(guild, params)

edit(guild(), Enum.t()) :: map()

Modifies a given guild.

Returns a guild object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - guild name

region - guild voice region

icon - base64 encoded 128x128 jpeg image

splash - base64 encoded 128x128 jpeg image

afk_timeout - voice AFK timeout in seconds

afk_channel_id - voice AFK channel

system_channel_id - channel to which system messages are sent

Refer to this for a broader explanation on the fields and their defaults.

edit_role_positions(guild, params)

edit_role_positions(guild(), list()) :: list() | map()

Modifies the positions of a set of roles on a given guild.

Returns a list of role objects upon success or a map containing error information.

Params

Must be a list of maps with the fields listed below.

id - snowflake of the role

position - sorting position of the role

get(id)

get(String.t()) :: map() | nil

Fetches a cached guild object.

Returns an object if found and nil otherwise.

get_ban(guild, user)

get_ban(guild(), String.t()) :: map()

Fetches the ban for a user on a given guild.

Returns a ban object upon success or a map containing error information.

get_bans(guild)

get_bans(guild()) :: list() | map()

Fetches the bans from a given guild.

Returns a list of ban objects or a map containing error information.

get_integrations(guild)

get_integrations(guild()) :: list() | map()

Fetches the integrations from a given guild.

Returns a list of integration objects or a map containing error information.

get_invites(guild)

get_invites(guild()) :: list() | map()

Fetches the invites from a given guild.

Returns a list of invite objects upon success or a map containing error information.

get_member(guild, user)

get_member(guild(), String.t()) :: map()

Fetches a member from a given guild.

Returns a member object upon success or a map containing error information.

get_prune(guild, query)

get_prune(guild(), Keyword.t()) :: map()

Gets the number of members that would be removed in a prune.

Returns a map with a pruned field or a map containing error information.

Query

Must be a keyword list with the fields listed below.

days - number of days to count prune for (1 or more)

get_regions()

get_regions() :: list() | map()

Fetches a list of voice regions.

Returns a list of voice region objects or a map containing error information.

get_regions(guild)

get_regions(guild()) :: list() | map()

Fetches the voice regions for a given guild.

Returns a list of voice region objects or a map containing error information.

get_roles(guild)

get_roles(guild()) :: list() | map()

Fetches the roles from a given guild.

Returns a list of role objects upon success or a map containing error information.

get_vanity_code(guild)

get_vanity_code(guild()) :: String.t() | map()

Fetches the vanity url code of a given guild.

Returns a string representing the code or a map containing error information.

get_webhooks(guild)

get_webhooks(guild()) :: list() | map()

Fetches the webhooks from a given guild.

Returns a list of webhook objects or a map containing error information.

leave(guild)

leave(guild()) :: :ok | map()

Leaves from a given guild.

Returns the atom :ok upon success or a map containing error information.

list_members(term, query \\ [])

list_members(guild(), Keyword.t()) :: list() | map()

Fetches the members from a given guild.

Returns a list of member objects upon success or a map containing error information.

Query

Must be a keyword list with the fields listed below.

limit - max number of members to return (1-1000)

after - the highest user id in the previous page

Refer to this for a broader explanation on the fields and their defaults.

select(pattern \\ %{})

select(map()) :: list()

Filters the cached guild objects.

Returns a list of matched objects.

set_afk_channel(guild, channel)

set_afk_channel(guild(), String.t()) :: map()

Changes the voice AFK channel of a given guild.

Returns a guild object upon success or a map containing error information.

set_afk_timeout(guild, timeout)

set_afk_timeout(guild(), Integer.t()) :: map()

Changes the voice AFK timeout of a given guild.

Returns a guild object upon success or a map containing error information.

set_icon(guild, icon)

set_icon(guild(), String.t()) :: map()

Changes the icon of a given guild.

Returns a guild object upon success or a map containing error information.

set_name(guild, name)

set_name(guild(), String.t()) :: map()

Changes the name of a given guild.

Returns a guild object upon success or a map containing error information.

set_region(guild, region)

set_region(guild(), String.t()) :: map()

Changes the region of a given guild.

Returns a guild object upon success or a map containing error information.

set_splash(guild, splash)

set_splash(guild(), String.t()) :: map()

Changes the splash of a given guild.

Returns a guild object upon success or a map containing error information.

set_system_channel(guild, channel)

set_system_channel(guild(), String.t()) :: map()

Changes the channel to which system messages are sent on a given guild.

Returns a guild object upon success or a map containing error information.

shard(guild)

shard(guild()) :: pid()

Used to grab the shard of a given guild.

Returns the PID of the shard’s process.

unban(guild, user)

unban(guild(), String.t()) :: :ok | map()

Removes the ban for a user on a given guild.

Returns the atom :ok upon success or a map containing error information.