coxir v0.1.0 Coxir.Struct.Integration

Defines methods used to interact with guild integrations.

Refer to this for a list of fields and a broader documentation.

Types

integration()

integration() :: String.t() | map()

Functions

delete(map)

delete(integration()) :: :ok | map()

Deletes a given integration.

Returns the atom :ok upon success or a map containing error information.

delete(integration, guild)

delete(String.t(), String.t()) :: :ok | map()

Deletes a given integration.

Refer to delete/1 for more information.

edit(map, params)

edit(integration(), Enum.t()) :: :ok | map()

Modifies a given integration.

Returns the atom :ok upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

expire_behavior - when an integration subscription lapses

expire_grace_period - period (in seconds) where lapsed subscriptions will be ignored

enable_emoticons - whether emoticons should be synced for this integration (Twitch)

Refer to this for a broader explanation on the fields and their defaults.

edit(integration, guild, params)

edit(String.t(), String.t(), Enum.t()) :: :ok | map()

Modifies a given integration.

Refer to edit/2 for more information.

sync(map)

sync(integration()) :: :ok | map()

Synchronizes a given integration.

Returns the atom :ok upon success or a map containing error information.

sync(integration, guild)

sync(String.t(), String.t()) :: :ok | map()

Synchronizes a given integration.

Refer to sync/1 for more information.