coxir v0.8.0 Coxir.Struct.Invite

Defines methods used to interact with guild invites.

Refer to this for a list of fields and a broader documentation.

Types

invite()

invite() :: String.t() | map()

Functions

accept(code)

accept(invite()) :: map()

Accepts a given invite.

Returns an invite object upon success or a map containing error information.

delete(code)

delete(invite()) :: map()

Deletes a given invite.

Returns an invite object upon success or a map containing error information.

get(id)

get(String.t()) :: map() | nil

Fetches an invite.

Returns an invite object upon success or a map containing error information.