coxir v0.8.0 Coxir.Struct.Member

Defines methods used to interact with guild members.

Refer to this for a list of fields and a broader documentation.

In addition, the following fields are also embedded.

user - a user object

voice - a voice channel object

roles - a list of role objects

Types

guild()

guild() :: map()

member()

member() :: map()

user()

user() :: map()

Functions

add_role(arg1, role)

add_role(member(), String.t()) :: :ok | map()

Adds a role to a given member.

Returns the atom :ok upon success or a map containing error information.

ban(arg1, query)

ban(member(), Keyword.t()) :: :ok | map()

Bans a given member.

Returns the atom :ok upon success or a map containing error information.

Query

Must be a keyword list with the fields listed below.

delete-message-days - number of days to delete the messages for (0-7)

reason - reason for the ban

edit(arg1, params)

edit(member(), Enum.t()) :: :ok | map()

Modifies a given member.

Returns the atom :ok upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

nick - value to set the member’s nickname to

roles - list of role ids the member is assigned

mute - whether the member is muted

deaf - whether the member is deafened

channel_id - id of a voice channel to move the member to

Refer to this for a broader explanation on the fields and their defaults.

get(server, member)

get(guild(), user()) :: map() | nil

Fetches a cached member object.

Returns an object if found and nil otherwise.

kick(arg1)

kick(member()) :: :ok | map()

Kicks a given member.

Returns the atom :ok upon success or a map containing error information.

move(member, channel)

move(member(), String.t()) :: :ok | map()

Changes the voice channel of a given member.

Returns the atom :ok upon success or a map containing error information.

remove_role(arg1, role)

remove_role(member(), String.t()) :: :ok | map()

Removes a role from a given member.

Returns the atom :ok upon success or a map containing error information.

select(pattern \\ %{})

select(map()) :: list()

Filters the cached member objects.

Returns a list of matched objects.

set_nick(tuple, name)

set_nick(member(), String.t()) :: map()

Changes the nickname of a given member.

Returns a map with a nick field or a map containing error information.