coxir v0.1.0 Coxir.Struct.Message

Defines methods used to interact with channel messages.

Refer to this for a list of fields and a broader documentation.

In addition, the following fields are also embedded.

guild - a guild object

channel - a channel object

Types

message()

message() :: String.t() | map()

Functions

delete(map)

delete(message()) :: :ok | map()

Deletes a given message.

Returns the atom :ok upon success or a map containing error information.

delete_all_reactions(map)

delete_all_reactions(message()) :: :ok | map()

Deletes all reactions from a given message.

Returns the atom :ok upon success or a map containing error information.

delete_reaction(map, emoji, user \\ "@me")

delete_reaction(message(), String.t(), String.t()) :: :ok | map()

Deletes a specific reaction from a given message.

Returns the atom :ok upon success or a map containing error information.

edit(map, content)

edit(message(), String.t() | Enum.t()) :: map()

Modifies a given message.

Returns a message object upon success or a map containing error information.

Content

Either a string or an enumerable with the fields listed below.

content - the message contents (up to 2000 characters)

embed - embedded rich content, refer to this

get(id)

get(String.t()) :: map() | nil

Fetches a cached message object.

Returns an object if found and nil otherwise.

get_reactions(map, emoji)

get_reactions(message(), String.t()) :: list() | map()

Fetches a list of users specific to a reaction on a given message.

Returns a list of user objects upon success or a map containing error information.

pin(map)

pin(message()) :: :ok | map()

Pins a given message.

Returns the atom :ok upon success or a map containing error information.

react(map, emoji)

react(message(), String.t()) :: :ok | map()

Reacts to a given message.

Returns the atom :ok upon success or a map containing error information.

reply(map, content)

reply(message(), String.t() | Enum.t()) :: map()

Replies to a given message.

Refer to Coxir.Struct.Channel.send_message/2 for more information.

select(pattern \\ %{})

select(map()) :: list()

Filters the cached message objects.

Returns a list of matched objects.

unpin(map)

unpin(message()) :: :ok | map()

Unpins a given message.

Returns the atom :ok upon success or a map containing error information.