coxir v0.8.0 Coxir.Struct.Role

Defines methods used to interact with guild roles.

Refer to this for a list of fields and a broader documentation.

Types

role()

role() :: String.t() | map()

Functions

delete(map)

delete(role()) :: :ok | map()

Deletes a given role.

Returns the atom :ok upon success or a map containing error information.

delete(role, guild)

delete(String.t(), String.t()) :: :ok | map()

Deletes a given role.

Refer to delete/1 for more information.

edit(map, params)

edit(role(), Enum.t()) :: map()

Modifies a given role.

Returns a role object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - name of the role

color - RGB color value

permissions - bitwise of the permissions

hoist - whether the role should be displayed separately

mentionable - whether the role should be mentionable

Refer to this for a broader explanation on the fields and their defaults.

edit(role, guild, params)

edit(String.t(), String.t(), Enum.t()) :: map()

Modifies a given role.

Refer to edit/2 for more information.

set_color(map, color)

set_color(role(), Integer.t()) :: map()

Changes the color of a given role.

Returns a role object upon success or a map containing error information.

set_color(role, guild, color)

set_color(String.t(), String.t(), Integer.t()) :: map()

Changes the color of a given role.

Refer to set_color/2 for more information.

set_hoist(map, bool)

set_hoist(role(), boolean()) :: map()

Changes the hoist flag of a given role.

Returns a role object upon success or a map containing error information.

set_hoist(role, guild, bool)

set_hoist(String.t(), String.t(), boolean()) :: map()

Changes the hoist flag of a given role.

Refer to set_hoist/2 for more information.

set_mentionable(map, bool)

set_mentionable(role(), boolean()) :: map()

Changes the mentionable flag of a given role.

Returns a role object upon success or a map containing error information.

set_mentionable(role, guild, bool)

set_mentionable(String.t(), String.t(), boolean()) :: map()

Changes the mentionable flag of a given role.

Refer to set_mentionable/2 for more information.

set_name(map, name)

set_name(role(), String.t()) :: map()

Changes the name of a given role.

Returns a role object upon success or a map containing error information.

set_name(role, guild, name)

set_name(String.t(), String.t(), String.t()) :: map()

Changes the name of a given role.

Refer to set_name/2 for more information.

set_permissions(map, permissions)

set_permissions(role(), Integer.t()) :: map()

Changes the permissions of a given role.

Returns a role object upon success or a map containing error information.

set_permissions(role, guild, permissions)

set_permissions(String.t(), String.t(), Integer.t()) :: map()

Changes the permissions of a given role.

Refer to set_permissions/2 for more information.