coxir v0.8.0 Coxir.Struct.User

Defines methods used to interact with Discord users.

Refer to this for a list of fields and a broader documentation.

In addition, the following fields are also embedded.

voice - a channel object

Types

user()

user() :: String.t() | map()

Functions

create_dm(recipient)

create_dm(user()) :: map()

Creates a DM channel with a given user.

Returns a channel object upon success or a map containing error information.

create_group(params)

create_group(Enum.t()) :: map()

Creates a group DM channel.

Returns a channel object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

access_tokens - access tokens of users

nicks - a map of user ids and their respective nicknames

Refer to this for a broader explanation on the fields and their defaults.

edit(params)

edit(Enum.t()) :: map()

Modifies the local user.

Returns an user object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

username - the user’s username

avatar - the user’s avatar

Refer to this for a broader explanation on the fields and their defaults.

get(id)

get(String.t()) :: map() | nil

Fetches a cached user object.

Returns an object if found and nil otherwise.

get_connections()

get_connections() :: list() | map()

Fetches a list of connections for the local user.

Refer to this for more information.

get_dms()

get_dms() :: list()

Fetches a list of DM channels for the local user.

Returns a list of channel objects upon success or a map containing error information.

get_guilds(query \\ [])

get_guilds(Keyword.t()) :: list() | map()

Fetches a list of guilds for the local user.

Returns a list of partial guild objects or a map containing error information.

Query

Must be a keyword list with the fields listed below.

before - get guilds before this guild ID

after - get guilds after this guild ID

max - max number of guilds to return

Refer to this for a broader explanation on the fields and their defaults.

get_id()

get_id() :: String.t()

Computes the local user’s ID.

Returns a snowflake.

select(pattern \\ %{})

select(map()) :: list()

Filters the cached user objects.

Returns a list of matched objects.

send_message(user, content)

send_message(user(), String.t() | Enum.t()) :: map()

Sends a direct message to a given user.

Refer to Coxir.Struct.Channel.send_message/2 for more information.

set_avatar(avatar)

set_avatar(String.t()) :: map()

Changes the avatar of the local user.

Returns an user object upon success or a map containing error information.

Avatar

Either a proper data URI scheme or the path of an image’s file.

Refer to this for a broader explanation.

set_username(name)

set_username(String.t()) :: map()

Changes the username of the local user.

Returns an user object upon success or a map containing error information.