coxir v0.8.0 Coxir.Struct.Webhook

Defines methods used to interact with channel webhooks.

Refer to this for a list of fields and a broader documentation.

Functions

delete(webhook)

delete(String.t()) :: :ok | map()

Deletes a given webhook.

Returns the atom :ok upon success or a map containing error information.

delete_with_token(webhook, token)

delete_with_token(String.t(), String.t()) :: :ok | map()

Deletes a given webhook.

Refer to this for more information.

edit(webhook, params)

edit(String.t(), Enum.t()) :: map()

Modifies a given webhook.

Returns a webhook object upon success or a map containing error information.

Params

Must be an enumerable with the fields listed below.

name - the default name of the webhook

avatar - image for the default webhook avatar

channel_id - the new channel id to be moved to

Refer to this for a broader explanation on the fields and their defaults.

edit_with_token(webhook, token, params)

edit_with_token(String.t(), String.t(), Enum.t()) :: map()

Modifies a given webhook.

Refer to this for more information.

execute(webhook, token, params, wait \\ false)

execute(String.t(), String.t(), Enum.t(), boolean()) :: map()

Executes a given webhook.

Refer to this for more information.

execute_github(webhook, token, params, wait \\ false)

execute_github(String.t(), String.t(), Enum.t(), boolean()) :: map()

Executes a given GitHub webhook.

Refer to this for more information.

execute_slack(webhook, token, params, wait \\ false)

execute_slack(String.t(), String.t(), Enum.t(), boolean()) :: map()

Executes a given Slack webhook.

Refer to this for more information.

get(id)

get(String.t()) :: map() | nil

Fetches a webhook.

Returns a webhook object upon success or a map containing error information.

get_with_token(webhook, token)

get_with_token(String.t(), String.t()) :: map()

Fetches a webhook.

Refer to this for more information.

set_avatar(webhook, avatar)

set_avatar(String.t(), String.t()) :: map()

Changes the avatar of a given webhook.

Returns a webhook object upon success or a map containing error information.

set_channel(webhook, channel)

set_channel(String.t(), String.t()) :: map()

Changes the channel of a given webhook.

Returns a webhook object upon success or a map containing error information.

set_name(webhook, name)

set_name(String.t(), String.t()) :: map()

Changes the name of a given webhook.

Returns a webhook object upon success or a map containing error information.