Coxir.Voice (coxir v2.0.0)

Allows for interaction with Discord voice channels.

Types

join_options()

Specs

join_options() :: [
  as: Coxir.Gateway.gateway(),
  self_deaf: boolean() | false,
  self_mute: boolean() | false
]

The options that must be passed to join/2.

leave_options()

Specs

leave_options() :: [{:as, Coxir.Gateway.gateway()}]

The options that must be passed to leave/2.

play_options()

Specs

play_options() ::
  [{:player, Coxir.Player.t() | Coxir.Player.Default}] | Coxir.Player.options()

The options that can be passed to play/3.

Functions

join(channel, options)

Specs

join(Coxir.Channel.t(), join_options()) :: Coxir.Voice.Instance.instance()

Joins a given voice channel.

If the user is already in the channel, the function acts as a no-op.

If the user is in a different channel of the same guild, it will stop playing and then switch.

leave(arg1, options)

Specs

leave(Coxir.Guild.t() | Coxir.Channel.t(), leave_options()) :: :ok

Leaves from a given voice channel, or the active voice channel for a guild.

pause(instance)

Specs

pause(Coxir.Voice.Instance.instance()) :: :ok | {:error, :no_player}

Pauses audio playback on a given instance.

play(instance, playable, options \\ [])

Specs

play(Coxir.Voice.Instance.instance(), Coxir.Player.playable(), play_options()) ::
  :ok | {:error, term()}

Begins playing audio on a given instance.

Refer to the documentation of the player in use for more information on the arguments.

If no custom player is provided, the default Coxir.Player.Default will be used.

playing?(instance)

Specs

playing?(Coxir.Voice.Instance.instance()) :: boolean()

Returns whether audio is currently playing on a given instance.

resume(instance)

Specs

resume(Coxir.Voice.Instance.instance()) :: :ok | {:error, :no_player}

Resumes audio playback on a given instance.

stop_playing(instance)

Specs

stop_playing(Coxir.Voice.Instance.instance()) :: :ok

Stops playing audio on a given instance.